feat: 강의 필터링 및 정렬 기능 고도화 및 통계 필드 캐싱 적용
- Specification을 활용한 동적 필터링 및 다양한 정렬 옵션(인기순, 평점순 등) 추가
- 조회 성능 최적화를 위해 Course 엔티티에 평균 평점과 수강생 수 캐싱 필드 도입
- 리뷰 작성/삭제 및 수강 신청/취소 시 통계 데이터가 자동으로 갱신되도록 로직 개선
- 평점 데이터 타입을 BigDecimal로 변경하여 정밀도 향상

+++ backend/DevClass/src/main/java/rlaxogh76/DevClass/domain/course/dto/CourseDetailResponse.java
@@ -2,6 +2,8 @@ package rlaxogh76.DevClass.domain.course.dto;
 
 import rlaxogh76.DevClass.domain.course.entity.Course;
 
+import java.math.BigDecimal;
+
 public record CourseDetailResponse(
         Long id,
         String title,
@@ -12,13 +14,13 @@ public record CourseDetailResponse(
         String category,
         String level,
         int lectureCount,
-        double averageRating,
+        BigDecimal averageRating,
         int reviewCount
 ) {
     public static CourseDetailResponse of(
             Course course,
             int lectureCount,
-            double averageRating,
+            BigDecimal averageRating,
             int reviewCount
     ) {
         return new CourseDetailResponse(
