+++ backend/DevClass/src/main/java/rlaxogh76/DevClass/domain/review/repository/ReviewRepository.java
@@ -5,13 +5,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
 
     @Query("SELECT COALESCE(AVG(r.rating), 0.0) FROM Review r WHERE r.course.id = :courseId")
-    double findAverageRatingByCourseId(@Param("courseId") Long courseId);
+    BigDecimal findAverageRatingByCourseId(@Param("courseId") Long courseId);
 
     List<Review> findByCourseIdOrderByCreatedAtDesc(Long courseId);
